requirements.txt
asgiref==3.2.7
Django==3.0.5
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-grappelli==2.14.1
django-js-asset==1.2.2
django-model-utils==1.3.1
gunicorn==20.0.4
Pillow==7.1.1
psycopg2-binary==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
